Change adequacy statement from Definition to Lemma in adequacy.v
It really confused me that it was a definition and not a lemma. After a bit of thinking, I realised that they aren't that different, so I thought there might be some subtle thing going on with Coq, but apparently not. Ralf just changed it in Iris' HeapLang so it's probably good to change it here as well https://gitlab.mpi-sws.org/iris/iris/-/commit/ecd139336fb207a52d8b71c5597e88861b8142c5